[rb] update print page specs to follow conventions
