+++ rb/spec/integration/selenium/webdriver/chrome/driver_spec.rb
@@ -52,6 +52,34 @@ module Selenium
             driver.execute_cdp('Page.removeScriptToEvaluateOnNewDocument', identifier: res['identifier'])
           end
         end
+
+        describe '#print_options' do
+          let(:magic_number) { 'JVBER' }
+          let(:options) { Chrome::Options.new(args: ['--headless']) }
+
+          it 'should return base64 for print command' do
+            create_driver!(capabilities: options) do |driver|
+              driver.navigate.to url_for('printPage.html')
+              expect(driver.print_page).to include(magic_number)
+            end
+          end
+
+          it 'should print with orientation' do
+            create_driver!(capabilities: options) do |driver|
+              driver.navigate.to url_for('printPage.html')
+              expect(driver.print_page(orientation: 'landscape')).to include(magic_number)
+            end
+          end
+
+          it 'should print with valid params' do
+            create_driver!(capabilities: options) do |driver|
+              driver.navigate.to url_for('printPage.html')
+              expect(driver.print_page(orientation: 'landscape',
+                                       page_ranges: ['1-2'],
+                                       page: {width: 30})).to include(magic_number)
+            end
+          end
+        end
       end
     end # Chrome
   end # WebDriver
